requirements.txt
rich
dnspython
requests
python-whois